test: convert test-cm-cli and test-evmu to lester spec format
Rename test-cm-cli.lua -> spec-cm-cli.lua and test-evmu.lua ->
spec-evmu.lua, converting both from stand-alone scripts to lester
modules (describe/it/expect.*) aggregated by test-spec.lua.

test-cm-cli and test-evmu Makefile targets are kept as dev-only
coverage-subset runners (filter on test-spec.lua); they are removed
from the test: aggregate, CI coverage job, and run-coverage-local.sh
because test-lua already covers them via test-spec.lua, avoiding
duplicate runs.

Replace tools/template/test-cm-cli.template with test-spec.template,
which sets LUA_PATH_5_4, LUA_CPATH_5_4, CM_CLI, and
CARTESI_IMAGES_PATH and installs as $(BIN_INSTALL_PATH)/test-spec.
