+++ tests/lua/test-spec.lua
@@ -29,6 +29,8 @@ require("spec-hash-tree")
 require("spec-collect-hashes")
 require("spec-backing-stores")
 require("spec-runtime-console")
+require("spec-cm-cli")
+require("spec-evmu")
 
 lester.report() -- Print overall statistic of the tests run.
 lester.exit() -- Exit with success if all tests passed.
