+++ tools/template/test-spec.template
@@ -2,4 +2,5 @@
 export LUA_PATH_5_4="ARG_LUA_PATH;${LUA_PATH_5_4:-;}"
 export LUA_CPATH_5_4="ARG_LUA_CPATH;${LUA_CPATH_5_4:-;}"
 export CM_CLI="${CM_CLI:-ARG_CM_CLI}"
-lua5.4 "ARG_LUA_RUNTIME_PATH/test-cm-cli.lua" "$@"
+export CARTESI_IMAGES_PATH="${CARTESI_IMAGES_PATH:-ARG_IMAGES_PATH}"
+lua5.4 "ARG_LUA_RUNTIME_PATH/test-spec.lua" "$@"
